working snake demo
