+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "rustnes"
 version = "0.1.0"
 authors = ["Parker Jones <parker.emailaddress@sharpspring.com>"]
-edition = "2018"
+edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -13,7 +13,8 @@ wrapping_arithmetic = "0.1.0"
 lazy_static = "1.4.0"
 bitflags = "1.2.1"
 sdl2 = "0.35.1"
-rand = "=0.8.4"
+rand = "=0.7.3"
 env_logger = "0.9"
 log = "0.4"
 dotenv = "0.15.0"
+anyhow = "1.0.75"
