+++ src/main.rs
@@ -1,63 +1,18 @@
-mod cpu;
-mod opcode;
-mod ops;
-mod util;
+pub mod cpu;
+pub mod util;
 
-#[macro_use]
-extern crate lazy_static;
-
-#[macro_use]
-extern crate bitflags;
-
-use crate::cpu::Mem;
-use crate::cpu::CPU;
-
-use log::debug;
+use cpu::CPU;
+use cpu::cpu::Mem;
 use rand::Rng;
 use sdl2::event::Event;
+use sdl2::EventPump;
 use sdl2::keyboard::Keycode;
 use sdl2::pixels::Color;
 use sdl2::pixels::PixelFormatEnum;
-use sdl2::EventPump;
-use dotenv::dotenv;
 use util::snake_game_code;
+// use std::time::Duration;
 
-fn handle_user_input(cpu: &mut CPU, event_pump: &mut EventPump) {
-    for event in event_pump.poll_iter() {
-        match event {
-            Event::Quit { .. }
-            | Event::KeyDown {
-                keycode: Some(Keycode::Escape),
-                ..
-            } => std::process::exit(0),
-            Event::KeyDown {
-                keycode: Some(Keycode::W),
-                ..
-            } => {
-                cpu.mem_write(0xff, 0x77);
-            }
-            Event::KeyDown {
-                keycode: Some(Keycode::S),
-                ..
-            } => {
-                cpu.mem_write(0xff, 0x73);
-            }
-            Event::KeyDown {
-                keycode: Some(Keycode::A),
-                ..
-            } => {
-                cpu.mem_write(0xff, 0x61);
-            }
-            Event::KeyDown {
-                keycode: Some(Keycode::D),
-                ..
-            } => {
-                cpu.mem_write(0xff, 0x64);
-            }
-            _ => { /* do nothing */ }
-        }
-    }
-}
+
 
 fn color(byte: u8) -> Color {
     match byte {
@@ -75,12 +30,10 @@ fn color(byte: u8) -> Color {
 
 fn read_screen_state(cpu: &CPU, frame: &mut [u8; 32 * 3 * 32]) -> bool {
     let mut frame_idx = 0;
-    let mut update = true;
-    // debug!("{:?}", frame);
+    let mut update = false;
     for i in 0x0200..0x600 {
         let color_idx = cpu.mem_read(i as u16);
         let (b1, b2, b3) = color(color_idx).rgb();
-        //debug!("{:#04X?} ({}, {}, {})", i, b1, b2, b3);
         if frame[frame_idx] != b1 || frame[frame_idx + 1] != b2 || frame[frame_idx + 2] != b3 {
             frame[frame_idx] = b1;
             frame[frame_idx + 1] = b2;
@@ -92,19 +45,37 @@ fn read_screen_state(cpu: &CPU, frame: &mut [u8; 32 * 3 * 32]) -> bool {
     update
 }
 
-fn main() {
-    //init logging
-    dotenv().ok();
-    env_logger::init();
+fn handle_user_input(cpu: &mut CPU, event_pump: &mut EventPump) {
+    for event in event_pump.poll_iter() {
+        match event {
+            Event::Quit { .. } | Event::KeyDown { keycode: Some(Keycode::Escape), .. } => {
+                std::process::exit(0)
+            },
+            Event::KeyDown { keycode: Some(Keycode::W), .. } => {
+                cpu.mem_write(0xff, 0x77);
+            },
+            Event::KeyDown { keycode: Some(Keycode::S), .. } => {
+                cpu.mem_write(0xff, 0x73);
+            },
+            Event::KeyDown { keycode: Some(Keycode::A), .. } => {
+                cpu.mem_write(0xff, 0x61);
+            },
+            Event::KeyDown { keycode: Some(Keycode::D), .. } => {
+                cpu.mem_write(0xff, 0x64);
+            }
+            _ => {/* do nothing */}
+        }
+    }
+}
 
+fn main() {
     // init sdl2
     let sdl_context = sdl2::init().unwrap();
     let video_subsystem = sdl_context.video().unwrap();
     let window = video_subsystem
         .window("Snake game", (32.0 * 10.0) as u32, (32.0 * 10.0) as u32)
         .position_centered()
-        .build()
-        .unwrap();
+        .build().unwrap();
 
     let mut canvas = window.into_canvas().present_vsync().build().unwrap();
     let mut event_pump = sdl_context.event_pump().unwrap();
@@ -112,27 +83,35 @@ fn main() {
 
     let creator = canvas.texture_creator();
     let mut texture = creator
-        .create_texture_target(PixelFormatEnum::RGB24, 32, 32)
-        .unwrap();
+        .create_texture_target(PixelFormatEnum::RGB24, 32, 32).unwrap();
+    
+
+    let game_code = snake_game_code();
 
 
+    //load the game
     let mut cpu = CPU::new();
-    cpu.load(snake_game_code());
+    cpu.load(game_code);
     cpu.reset();
+
     let mut screen_state = [0 as u8; 32 * 3 * 32];
     let mut rng = rand::thread_rng();
-    cpu.mem_write(0xfe, rng.gen_range(1..16));
+
+    // run the game cycle
     cpu.run_with_callback(move |cpu| {
         handle_user_input(cpu, &mut event_pump);
 
-        let result = read_screen_state(cpu, &mut screen_state); 
-        debug!("read screen state {}", result);
-        if result {
+        cpu.mem_write(0xfe, rng.gen_range(1, 16));
+
+        if read_screen_state(cpu, &mut screen_state) {
             texture.update(None, &screen_state, 32 * 3).unwrap();
+
             canvas.copy(&texture, None, None).unwrap();
+
             canvas.present();
         }
 
-        ::std::thread::sleep(std::time::Duration::new(0, 70_000));
+        ::std::thread::sleep(std::time::Duration::new(0, 60_000));
     });
+
 }
