+++ src/util.rs
@@ -1,15 +1,112 @@
+// use std::fs::File;
+// use std::io::Read;
+
+use sdl2::EventPump;
+use sdl2::event::Event;
+use sdl2::keyboard::Keycode;
+use sdl2::pixels::Color;
+
+use crate::cpu::cpu::CPU;
+use crate::cpu::cpu::Mem;
 use std::fs::File;
 use std::io::Read;
-
-pub fn load_bytes_from_file(path: &str) -> std::io::Result<Vec<u8>> {
-    let mut file = File::open(path)?;
-    let mut buffer = Vec::new();
-    file.read_to_end(&mut buffer)?;
-    Ok(buffer)
-}
+use std::path::Path;
+//ain't working
+// pub fn load_bytes_from_file(path: &str) -> std::io::Result<Vec<u8>> {
+//     let mut file = File::open(path)?;
+//     let mut buffer = Vec::new();
+//     file.read_to_end(&mut buffer)?;
+//     Ok(buffer)
+// }
 
 pub fn snake_game_code() -> Vec<u8> {
-    let mut snake_game = load_bytes_from_file("roms/snake.ch8").unwrap();
-    snake_game.append(&mut vec![0; 4096 - snake_game.len()]);
-    snake_game
+    load_ch8("roms/snake.ch8").expect("failed to load file")
+}
+
+pub fn color(byte: u8) -> Color {
+    match byte {
+        0 => sdl2::pixels::Color::BLACK,
+        1 => sdl2::pixels::Color::WHITE,
+        2 | 9 => sdl2::pixels::Color::GREY,
+        3 | 10 => sdl2::pixels::Color::RED,
+        4 | 11 => sdl2::pixels::Color::GREEN,
+        5 | 12 => sdl2::pixels::Color::BLUE,
+        6 | 13 => sdl2::pixels::Color::MAGENTA,
+        7 | 14 => sdl2::pixels::Color::YELLOW,
+        _ => sdl2::pixels::Color::CYAN,
+    }
+}
+
+pub fn handle_user_input(cpu: &mut CPU, event_pump: &mut EventPump) {
+    for event in event_pump.poll_iter() {
+        match event {
+            Event::Quit { .. }
+            | Event::KeyDown {
+                keycode: Some(Keycode::Escape),
+                ..
+            } => std::process::exit(0),
+            Event::KeyDown {
+                keycode: Some(Keycode::W),
+                ..
+            } => {
+                cpu.mem_write(0xff, 0x77);
+            }
+            Event::KeyDown {
+                keycode: Some(Keycode::S),
+                ..
+            } => {
+                cpu.mem_write(0xff, 0x73);
+            }
+            Event::KeyDown {
+                keycode: Some(Keycode::A),
+                ..
+            } => {
+                cpu.mem_write(0xff, 0x61);
+            }
+            Event::KeyDown {
+                keycode: Some(Keycode::D),
+                ..
+            } => {
+                cpu.mem_write(0xff, 0x64);
+            }
+            _ => { /* do nothing */ }
+        }
+    }
+}
+
+
+pub fn read_screen_state(cpu: &CPU, frame: &mut [u8; 32 * 3 * 32]) -> bool {
+    let mut frame_idx = 0;
+    let mut update = true;
+    // log::debug!("{:?}", frame);
+    for i in 0x0200..0x600 {
+        let color_idx = cpu.mem_read(i as u16);
+        let (b1, b2, b3) = color(color_idx).rgb();
+        //debug!("{:#04X?} ({}, {}, {})", i, b1, b2, b3);
+        if frame[frame_idx] != b1 || frame[frame_idx + 1] != b2 || frame[frame_idx + 2] != b3 {
+            frame[frame_idx] = b1;
+            frame[frame_idx + 1] = b2;
+            frame[frame_idx + 2] = b3;
+            update = true;
+        }
+        frame_idx += 3;
+    }
+    update
+}
+
+
+fn load_ch8(path: &str) -> anyhow::Result<Vec<u8>> {
+    let path = Path::new(path);
+    let mut file = File::open(&path)?;
+    let mut contents = String::new();
+    file.read_to_string(&mut contents)?;
+
+    parse_hex_string(&contents).map_err(Into::into)
+}
+
+fn parse_hex_string(hex_string: &str) -> Result<Vec<u8>, std::num::ParseIntError> {
+    hex_string.split(",")  // Split the string by commas
+        .map(|s| s.trim()) // Trim whitespace
+        .map(|s| u8::from_str_radix(&s[2..], 16)) // Parse each hex value
+        .collect() // Collect into a Vec<Result<u8, ParseIntError>>
 }
